Exclude escape characters. Add abstract text and input, with chunking.
The results of parsing now exclude escape characters, so the results
represent logical strings of characters without any escapes.  This is
achieved using new types for concatenating chunks of text that break
around and exclude escape characters.  This is zero-copy when the input
source and chunk types are zero-copy, which is true for the provided
types for using `&str`, and users can implement other types that are
zero-copy.

In addition, the AST and parsing now support fully generic text and
input types with chunking, which enables any source of characters, from
any representation, to be used, including streaming sources and ropes
from other libraries.  This includes support for generic character
position information provided by sources and reflected in the parsed
results and errors.

The only limitation is due to the lack of the upcoming
`generic_associated_types` feature of Rust, but this is not very
limiting due to our involved design that is able to generically relate
lifetimes of borrows needed for our iterators for many, but not all,
possible text types.  Once the GAT feature becomes stable, it probably
should be used instead or in conjunction because it should enable all
possible types.  While the current limited design is somewhat complex,
it does provide one generic iteration implementation for all potential
text types, which eases integration of them.

In the core crate, a text type is provided that reuses a parser's
generic datum allocator for the dynamic allocation needed to break
chunks around escape characters.  This enables `no_std` applications to
still have this ability.  In the full crate, a text type is provided
that uses heap allocation, which is simpler to use.  Both of these text
types work with any generic chunk types, enabling flexibility.  A
zero-copy chunk type for `&str` sources is provided and it also has
positional information.

The main Parser is now a struct parameterized over separate types that
provide its separate functionality.  This assists the new design and
allows more modular reuse of the separated functionality, which the
tests now take advantage of and which should be good for the future.

Unfortunately, the internal parse recursion now has a bug, caused by the
new excluding of escape characters, where escaping is being applied
multiple times when it shouldn't be.  This must be fixed next and will
require redesigning this recursion.

The existing tests are refactored but kept basically the same.  Except,
test suite #0 no longer checks character position info because the new
design allows sources that do not supply it or supply it as generic
arbitrary types, and we want that suite to be usable with such sources.

Deleted core/tests/refcell_stack_alloc.rs because it'd now require
having its own text type implementation, and, since the RefCell use was
always dubious, I don't want to maintain it anymore.

Many new tests need to be added for the new design's parts, but this
should wait until reorganizing into separate modules is done.

More clean-up is needed, but this commit contains the fleshed-out and
working new design.

This commit squashes into one the changes of many commits of messy,
churning, inconsistent experiments in other branches that led to this.
Those would be too confusing to include in the main history.