Add premade generic Datum allocators that use boxed or slice types.
Refactored the integration tests to use these equivalently.

TextConcat's DatumAllocator type parameter, and TextVec's impl of it,
are no longer bounded because that caused an error with mutually
recursive cycles of resolution when used with DatumAllocator impls with
generic text types.  It is fine to have it be unbounded, for impls that
ignore the parameter anyway, which is most of them.  Types like
TextDatumList that do use the parameter can still place their own bounds
on it.

DatumAllocator's Text associated type is no longer bounded by TextConcat
because that caused an error with mutually recursive cycles of
resolution when used with TextConcat impls like TextDatumList that need
to place bounds on TextConcat's DatumAllocator type parameter.

SourceStream now only has a single type parameter, for the
DatumAllocator, because the other old one for the text type was not
really needed and it would've required changing to not be bounded by
TextConcat to fit with the above changes.  Removing the other avoids
that and makes it simpler.