Add premade generic OperatorBindings that use HashMap or slice types.
The Hash trait is now derived for Datum and for all premade Datum
reference types, and implemented for all premade Text types, so they can
be used in Datums used in HashMapOperatorBindings.

The OperatorBindings::lookup method now returns borrowed references to
the Combiners (instead of owned values) because this enables
single-ownership types like Box to be used to hold the functions (as dyn
trait objects) while remaining owned by the OperatorBindings instance
they belong to.  Also, the lookup method now borrows `self` immutably
(instead of mutably) as part of maintaining the single-lookup efficiency
mentioned below.  This now precludes dynamically mutating the bindings
when lookup is called, but such a use was not really motivated and the
loss seems worth it and interior mutability should still be usable in
such cases.

The type of combiner/macro functions is now `dyn Fn` (instead of `dyn
FnMut`), because mutably borrowing the Combiner values (as was needed to
call their functions) caused conflicts in the parser logic that needs to
be able to have multiple borrows of them within the same lifetime to
achieve recursive parsing (without resorting to inefficient approaches
like doing operator binding lookups twice).  This now precludes mutable
closures as functions, but this seems worth it to maintain single-lookup
efficiency because needing such closures seems unlikely and interior
mutability should still be usable in such cases.

The parsing methods of ParseIter now take as arguments borrows of the
individual ParseIter and Parser fields, because this is needed to avoid
interprocedural borrow conflict errors that happened due to method calls
mutably borrowing `self` during the lifetime of borrows of Combiner
values.  Separating it into borrows of the individual fields allows the
logic to use some mutably and others immutably without conflict across
various method calls which enables borrowing (immutably) the Combiner
values across the extent of mutable use of other fields.  This is part
of maintaining the single-lookup efficiency mentioned above.

Test suite #0 now requires an implementation of the new
TestOperatorBindings (instead of EmptyOperatorBindings) to enable tests
to choose an implementation of OperatorBindings they want to be tested,
and the suite's section for Combiner bindings is refactored for this.
The integration tests that are not about testing an OperatorBindings now
use the new BasicTestOperatorBindings instead, to work with this new
requirement of the suite.  The new premade OperatorBindings
implementations have premade implementations of TestOperatorBindings
which are used by their integration tests.