Add premade Parsers and functions for common basic parsing applications.
These make using the parser much easier (than needing to be familiar
with the large amount of parameterization in the crates).  One for
parsing inputs that are single in-memory strings, and another for
parsing inputs that are chunks from streaming sources, both using
standard Box and HashMap.  These set pre-chosen types for most of the
generic parameters, and have "nothing" defaults for the type parameters
for Datum::Extra and Error::FailedCombiner.