Reduce default tree size used by the tests to 2^21.
This reduces the amount of RAM used to 2.13G when running the tests with
--test-threads=4, on my 4-core 64-bit machine, compared to 17G with the
size set to 2^24.  A size of 2^20 is more than enough to blow the stack
(on my system), with our special Drop impls commented-out, so 2^21
should be sufficient to ensure that our special Drop impls are being
properly tested as avoiding stack overflows.