Add macro that provides single definition of lints across workspace.
Now only a single file needs to be maintained for the crates to use the same
lints.  Replaces the lint attributes that were repeated at the roots of the
crates.

Add kul_lints crate that has declare_lints_on_top_module macro.  This crate
enables using this macro across the other crates.  Because macros cannot expand
to inner attributes, this macro must expand to outer attributes for the lints
and these must apply to a following module item so the lints will apply
throughout.  As such, move the contents of each lib.rs to new top.rs and change
each lib.rs to just call this macro that applies the lints to a private top
module and publicly re-exports its items.

This preserves the same logical structures of the module/item hierarchies and so
preserves the same API, and all tests pass without modification.  The generated
documentation is the same (the doc-comments of each lib.rs are kept in those
root modules, to preserve this).

Some items now require pub(crate) because they were private in a root module but
are now in a top module and must remain accessible throughout a crate.

The version patch number is incremented, to reflect these changes.