Restore the lib.rs files as the "top" (but not root) of the module hierarchies.
This enables building without the lints, by simply commenting-out the
`[lib] path` in the Cargo.toml files, which is better organized for potentially
removing kul_lints in the future.

Now the root modules of the crates are named lib-lints.rs, which better reflects
their purpose.

Commit 5554f9e had moved lib.rs to top.rs and added new lib.rs.  Now, those
lib.rs are moved to lib-lints.rs and those top.rs are moved back to lib.rs.